WIP check-in
Thsee changes mostly bring the Vault V2 project back in line with the develop branch.  There are still many companies that render under ITR_UI but not ITR_DV.  And two that do render--AEP and Enel--are pretty wrong (one way too high, the other way too low).  But this should be a good checkpoint from which further progress can be made.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>